requirements.txt
asgiref==3.9.1
colorama==0.4.6
Django==5.2.5
django-cors-headers==4.7.0
django-formtools==2.5.1
django-otp==1.6.1
django-phonenumber-field==8.1.0
django-two-factor-auth==1.17.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
phonenumbers==9.0.11
PyJWT==2.10.1
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
